Adding GetPreSign S3 function
Enables for anyone to reach a object without using any key temporarly.
A key in this case is the name of the object that you want to enable for
anyone to download without any login.

I don't think this is anything that we will use for this project but
still good to know about.